requirements.txt
openmeteo_requests
requests_cache
retry_requests
mysql-connector-python
timezonefinder
pytz
pandas
